Reintegrate new WebSocketClient
